Skf to png

+++ world/LoopScript.h
@@ -75,7 +75,7 @@ Foundation, Inc., 59 Temple Place - Suite 330, Boston, MA  02111-1307, USA.
 
 #define LOOPSCRIPT(name,tokens) const uint8 name[] = { tokens, LS_TOKEN_END }
 
-#define LS_CONSTANT(val)			((val)&0xFF), ((val>>8)&0xFF)
+#define LS_CONSTANT(val)			static_cast<uint8>((val)&0xFF), static_cast<uint8>((val>>8)&0xFF)
 #define LS_INT(val)					LS_TOKEN_INT, LS_CONSTANT(val)
 
 #define LS_OP(left, op, right)		left, right, op
